Fixed a query in the ObsProjectDaoImpl Fixed some DAO to be used with ACS-9.0
git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/SCHEDULING@141006 523d945c-050c-4681-91ec-863ad3bb968a

+++ Common/src/alma/scheduling/datamodel/obsproject/dao/SchedBlockDaoImpl.java
@@ -231,7 +231,7 @@ public class SchedBlockDaoImpl extends GenericDaoImpl implements SchedBlockDao {
 	public int countAll() {
         Query query = null;
         query = getSession().createQuery("select count(x) from SchedBlock x ");
-        return (Integer)query.uniqueResult();
+        return ((Long)query.uniqueResult()).intValue();
 	}
 
     @Override
